"[FAB2023-1177]feat: 화면 개발 store 이전"
